RM-69779 Support DC and CAS Mgr in Private Subnet
In order to comply with best security practices we need to deploy the
CAS Manager and Domain Controller in private subnets.  The dc module
cannot easily be modified to accomplish this, and has been absorbed into
the third_party_modules/teradici/editshare_deployments folder inside
terraform-deploy. The manager and connector can easily be modified by
eliminating the hardcoding of assigning a public IP to the instance,
and allowing the keypair used by the main terraform root module to be
applied to all of the teradici components.

References RM-69779