RM-71386 Adding Name tag to CAS root devices
Adding a Name tag to the root_block_device for the Connector and Manager
to support tracking resources.  To minimize the changes to the original
code from teradici, only setting the Name tag and not merging with the
common_tags as is done elsewhere, which means we can avoid a lot of
changes to pass that variable down into those modules.

References RM-71386